contrib/tf_serving/requirements.txt
Keras==2.2.*
Pillow==5.3.*
tensorflow-serving-api==1.12.*
